+++ src/main/kotlin/com/terraformation/backend/seedbank/api/AccessionSearchController.kt
@@ -1,0 +1,163 @@
+package com.terraformation.backend.seedbank.api
+
+import com.opencsv.CSVWriter
+import com.terraformation.backend.api.SeedBankAppEndpoint
+import com.terraformation.backend.db.FacilityId
+import com.terraformation.backend.search.SearchFieldPath
+import com.terraformation.backend.search.SearchResults
+import com.terraformation.backend.search.api.HasSearchNode
+import com.terraformation.backend.search.api.HasSortOrder
+import com.terraformation.backend.search.api.SearchFilter
+import com.terraformation.backend.search.api.SearchNodePayload
+import com.terraformation.backend.search.api.SearchResponsePayload
+import com.terraformation.backend.search.api.SearchSortOrderElement
+import com.terraformation.backend.seedbank.search.AccessionSearchService
+import io.swagger.v3.oas.annotations.Operation
+import io.swagger.v3.oas.annotations.media.Content
+import io.swagger.v3.oas.annotations.media.ExampleObject
+import io.swagger.v3.oas.annotations.media.Schema
+import io.swagger.v3.oas.annotations.responses.ApiResponse
+import java.io.ByteArrayOutputStream
+import java.io.OutputStreamWriter
+import java.nio.charset.StandardCharsets
+import java.time.Clock
+import java.time.LocalDateTime
+import java.time.format.DateTimeFormatter
+import javax.validation.constraints.NotEmpty
+import javax.ws.rs.BadRequestException
+import org.springframework.http.ContentDisposition
+import org.springframework.http.HttpHeaders
+import org.springframework.http.HttpStatus
+import org.springframework.http.ResponseEntity
+import org.springframework.web.bind.annotation.PostMapping
+import org.springframework.web.bind.annotation.RequestBody
+import org.springframework.web.bind.annotation.RequestMapping
+import org.springframework.web.bind.annotation.RestController
+
+@RequestMapping("/api/v1/seedbank/search")
+@RestController
+@SeedBankAppEndpoint
+class AccessionSearchController(
+    private val clock: Clock,
+    private val accessionSearchService: AccessionSearchService
+) {
+  @Operation(summary = "Searches for accessions based on filter criteria.")
+  @PostMapping
+  fun searchAccessions(
+      @RequestBody
+      @io.swagger.v3.oas.annotations.parameters.RequestBody(
+          content =
+              [
+                  Content(
+                      examples =
+                          [
+                              ExampleObject(
+                                  name = "example1",
+                                  description =
+                                      "Search for all accessions of species \"Species Name\" " +
+                                          "whose remaining seeds weigh between 100 and 200 " +
+                                          "milligrams or that have between 30 and 40 seeds " +
+                                          "remaining.",
+                                  // This value is parsed by the schema generator and rendered
+                                  // as JSON or YAML, so its formatting is irrelevant.
+                                  value =
+                                      """{ "fields": ["accessionNumber", "remainingQuantity", "remainingUnits"],
+                                           "search": {
+                                             "operation": "and", "children": [
+                                               { "operation": "field", "field": "species", "values": ["Species Name"] },
+                                               { "operation": "or", "children": [
+                                                   { "operation": "field", "field": "remainingGrams", "type": "Range", "values": ["100 Milligrams", "200 Milligrams"] },
+                                                   { "operation": "and", "children": [
+                                                       { "operation": "field", "field": "remainingUnits", "values": ["Seeds"] },
+                                                       { "operation": "field", "field": "remainingQuantity", "type": "Range", "values": ["30", "40"] } ] } ] } ] } }""")])])
+      payload: SearchAccessionsRequestPayload
+  ): SearchResponsePayload {
+    return SearchResponsePayload(
+        accessionSearchService.search(
+            payload.facilityId,
+            payload.fields,
+            payload.toSearchNode(),
+            payload.searchSortFields ?: emptyList(),
+            payload.cursor,
+            payload.count))
+  }
+
+  @ApiResponse(
+      responseCode = "200",
+      description = "Export succeeded.",
+      content =
+          [Content(mediaType = "text/csv", schema = Schema(type = "string", format = "binary"))])
+  @Operation(summary = "Exports the results of a search as a downloadable CSV file.")
+  @PostMapping("/export", produces = ["text/csv"])
+  fun exportAccessions(
+      @RequestBody payload: ExportAccessionsRequestPayload
+  ): ResponseEntity<ByteArray> {
+    if (payload.fields.any { it.isNested }) {
+      throw BadRequestException("Nested fields are not supported for CSV export.")
+    }
+
+    val searchResults =
+        accessionSearchService.search(
+            payload.facilityId,
+            payload.fields,
+            payload.toSearchNode(),
+            payload.searchSortFields ?: emptyList())
+    return exportCsv(payload, searchResults)
+  }
+
+  private fun exportCsv(
+      payload: ExportAccessionsRequestPayload,
+      searchResults: SearchResults
+  ): ResponseEntity<ByteArray> {
+    val dateAndTime =
+        DateTimeFormatter.ofPattern("yyyyMMdd-HHmmss").format(LocalDateTime.now(clock))
+    val filename = "seedbank-$dateAndTime.csv"
+    val byteArrayOutputStream = ByteArrayOutputStream()
+
+    // Write a UTF-8 BOM so Excel won't screw up the character encoding if there are non-ASCII
+    // characters.
+    byteArrayOutputStream.write(239)
+    byteArrayOutputStream.write(187)
+    byteArrayOutputStream.write(191)
+
+    CSVWriter(OutputStreamWriter(byteArrayOutputStream, StandardCharsets.UTF_8)).use { csvWriter ->
+      val header = payload.fields.map { it.searchField.displayName }.toTypedArray()
+      val fieldNames = payload.fields.map { "$it" }
+      csvWriter.writeNext(header, false)
+
+      searchResults.results.forEach { result ->
+        val values = fieldNames.map { fieldName -> result[fieldName]?.toString() }.toTypedArray()
+        csvWriter.writeNext(values, false)
+      }
+    }
+
+    val value = byteArrayOutputStream.toByteArray()
+    val headers = HttpHeaders()
+    headers.contentLength = value.size.toLong()
+    headers["Content-type"] = "text/csv;charset=UTF-8"
+    headers.contentDisposition = ContentDisposition.attachment().filename(filename).build()
+
+    return ResponseEntity(value, headers, HttpStatus.OK)
+  }
+}
+
+data class SearchAccessionsRequestPayload(
+    val facilityId: FacilityId,
+    @NotEmpty val fields: List<SearchFieldPath>,
+    override val sortOrder: List<SearchSortOrderElement>? = null,
+    override val filters: List<SearchFilter>? = null,
+    override val search: SearchNodePayload? = null,
+    val cursor: String? = null,
+    @Schema(
+        defaultValue = "10",
+    )
+    val count: Int = 10
+) : HasSearchNode, HasSortOrder
+
+data class ExportAccessionsRequestPayload(
+    val facilityId: FacilityId,
+    @NotEmpty val fields: List<SearchFieldPath>,
+    override val sortOrder: List<SearchSortOrderElement>? = null,
+    override val filters: List<SearchFilter>? = null,
+    override val search: SearchNodePayload? = null,
+) : HasSearchNode, HasSortOrder
